Organize contract and version folders
