+++ libs/contract/src/index.ts
@@ -1,16 +1,16 @@
 // State
-export * from './lib/+state/contract.model';
-export * from './lib/+state/contract.query';
-export * from './lib/+state/contract.service';
-export * from './lib/+state/contract.store';
-export * from './lib/+state/contract.firestore';
+export * from './lib/contract/+state/contract.model';
+export * from './lib/contract/+state/contract.query';
+export * from './lib/contract/+state/contract.service';
+export * from './lib/contract/+state/contract.store';
+export * from './lib/contract/+state/contract.firestore';
 
 // Modules
-export * from './lib/forms/party-name/party-name.module';
+export * from './lib/contract/forms/party-name/party-name.module';
 
 // Forms
-export * from './lib/forms/contract.form'
+export * from './lib/contract/forms/contract.form'
 
 // Guards
-export * from './lib/guards/active-contract.guard';
-export * from './lib/guards/active-organization-contracts.guard';
+export * from './lib/contract/guards/active-contract.guard';
+export * from './lib/contract/guards/active-organization-contracts.guard';
