+++ libs/contract/src/lib/contract/+state/contract.query.ts
