+++ libs/contract/src/lib/contract/forms/party-name/party-name.component.scss
